requirements.txt
djangorestframework==3.11.0
django-rest-swagger==2.2.0
django-filter==2.3.0
decorator=4.4.2
Django==3.2.3 
PyYAML==5.4.1
psycopg2==2.9.1
coreapi==2.3.1
coreschema==0.0.4
markdown==3.2.2;
pygments>=2.7.1,<2.8